+++ examples/benchmark_metric.py
@@ -1,7 +1,0 @@
-from aim import track
-import aim
-
-aim.init(True)
-
-for i in range(100000):
-    track(aim.metric, 'metric', i)