+++ main_de.py
@@ -1,0 +1,5 @@
+from aim.cli.de.commands import de_entry_point
+
+
+if __name__ == '__main__':
+    de_entry_point()