+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+## 2.0.0 Jun 18 2020
 - Tidy up aim and remove some artifacts (gorarakelyan)
 - Update AimContainerCMD to open connection on custom port (gorarakelyan)
 - Save passed process uuid to commit configs (gorarakelyan)